Ajustes para menu paciente
